Quick fixes
Proper value range for Rotation Angle/RadiusMultiplier, and restore
backwards compatibility for levels saved with version<1.2 libraries.

+++ LayerPropertiesForm.cs
@@ -218,8 +218,8 @@ namespace MLLE
                 DataSource.Hidden = Hidden.Checked;
                 DataSource.SpriteMode = (byte)SpriteMode.SelectedIndex;
                 DataSource.SpriteParam = (byte)SpriteParam.Value;
-                DataSource.RotationAngle = (byte)RotationAngle.Value;
-                DataSource.RotationRadiusMultiplier = (byte)RotationRadiusMultiplier.Value;
+                DataSource.RotationAngle = (int)RotationAngle.Value;
+                DataSource.RotationRadiusMultiplier = (int)RotationRadiusMultiplier.Value;
                 SourceForm.LevelHasBeenModified = true;
             }
